JSHintBear: Move to js package
