+++ bears/js/JSHintBear.py
