+++ bears/tests/js/test_files/jshint_config.js
