+++ bears/tests/js/test_files/jshint_test1.js
