+++ bears/tests/js/test_files/jshint_test2.js
